# public/swagger.json
{
	"components": {
...
		"schemas": {
			"SaveUpdateResItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"item_name": {
						"type": "string"
					},
					"item_price": {
						"type": "string"
					},
					"item_size": {
						"type": "string"
					},
					"item_type": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"item_name",
					"item_price",
					"item_size",
					"item_type",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetItem": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"getItemPrice": {
				"properties": {
					"item_price": {
						"type": "string"
					}
				},
				"required": [
					"item_price"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SaveUpdateResOrder": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"tableNo": {
						"type": "string"
					},
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"waiter": {
						"type": "string"
					},
					"o_status": {
						"type": "string"
					},
					"o_discription": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"tableNo",
					"item",
					"waiter",
					"o_status",
					"o_discription",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqOrder": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"tableNo": {
						"type": "string"
					},
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"o_discription": {
						"type": "string"
					}
				},
				"required": [
					"customerName",
					"tableNo",
					"item",
					"o_discription"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"customerName": {
						"type": "string"
					},
					"tableNo": {
						"type": "string"
					},
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"waiter": {
						"type": "string"
					},
					"o_discription": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"customerName",
					"tableNo",
					"item",
					"waiter",
					"o_discription"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SearchReqOrder": {
				"properties": {
					"item": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"item"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrderByStatus": {
				"properties": {
					"o_status": {
						"type": "string"
					}
				},
				"required": [
					"o_status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResWaiter": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"waiter_name": {
						"type": "string"
					},
					"waiter_email": {
						"type": "string"
					},
					"waiter_password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"waiter_name",
					"waiter_email",
					"waiter_password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetWaiter": {
				"properties": {
					"waiter_email": {
						"type": "string"
					},
					"waiter_password": {
						"type": "string"
					}
				},
				"required": [
					"waiter_email",
					"waiter_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqWaiter": {
				"properties": {
					"waiter_name": {
						"type": "string"
					},
					"waiter_email": {
						"type": "string"
					},
					"waiter_password": {
						"type": "string"
					}
				},
				"required": [
					"waiter_name",
					"waiter_email",
					"waiter_password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "http",
				"name": "Authorization",
				"in": "header",
				"scheme": "Bearer"
			}
		}
	},
	"info": {
		"title": "hotel-management-system",
		"version": "1.0.0",
		"description": "Hotel Management System by - Muhammad Ahsan",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/Items/getItemByID": {
			"post": {
				"operationId": "GetItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResItem"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetItem"
							}
						}
					}
				}
			}
		},
		"/Items/getMenu": {
			"post": {
				"operationId": "GetItemList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the items => anyone is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/Items/getItemsbyPrice": {
			"post": {
				"operationId": "GetItemByPrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the items with spacific price=> anyone is allowed",
				"tags": [
					"PMS - Items"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/getItemPrice"
							}
						}
					}
				}
			}
		},
...
		"/orders/getorderById": {
			"post": {
				"operationId": "Getorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"summary": "Get Single Order by order id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrder"
							}
						}
					}
				}
			}
		},
		"/orders/PlaceNewOrder": {
			"post": {
				"operationId": "Saveorder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"summary": "Take order from the customer => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqOrder"
							}
						}
					}
				}
			}
		},
		"/orders/getMyOrderlistByWaiterId": {
			"post": {
				"operationId": "GetorderListByWaiterId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get All the orders taken by current waiter => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/orders/updateorder": {
			"put": {
				"operationId": "UpdateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"summary": "Update Order by order Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqOrder"
							}
						}
					}
				}
			}
		},
...
		"/orders/getorderlist": {
			"post": {
				"operationId": "GetorderList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get list of all the orders => Only Login admin is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/orders/getorderbyItemID": {
			"post": {
				"operationId": "SearchorderByproduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get all the orders contain certain item => Only Login admin is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchReqOrder"
							}
						}
					}
				}
			}
		},
		"/orders/getOrderByStatus": {
			"post": {
				"operationId": "GetorderByStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get Single Item By Item Id => Only Login Waiter is allowed",
				"tags": [
					"PMS - Orders"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrderByStatus"
							}
						}
					}
				}
			}
		},
		"/waiter/loginwaiter": {
			"post": {
				"operationId": "Authwaiter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResWaiter"
								}
							}
						}
					}
				},
				"summary": "Log into the waiter account, which is registered by admin",
				"tags": [
					"PMS - Waiter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetWaiter"
							}
						}
					}
				}
			}
		},
		"/waiter/Registerwaiter": {
			"post": {
				"operationId": "SaveWaiter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResWaiter"
								}
							}
						}
					}
				},
				"summary": "Register new waiter => Only Login admin is allowed",
				"tags": [
					"PMS - Waiter"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqWaiter"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}